Filter Module/requirements.txt
pytest
pytest-html
selenium
undetected-chromedriver
pyyaml
